edit delete & changeBizStatus user fixed

+++ server/users/routes/usersRestController.js
@@ -112,21 +112,39 @@ router.put("/:id", auth, async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
-    const user = await changeUserBusinessStatus(id);
-    return res.send(user);
+    const user = req.user;
+
+    if (id !== user._id && !user.isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must bean an admin type user or the registered user to update its business status"
+      );
+
+    const changedStatusUser = await changeUserBusinessStatus(id);
+    return res.send(changedStatusUser);
   } catch (error) {
     return handleError(res, error.status || 500, error.message);
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
-    const user = await deleteUser(id);
-    return res.send(user);
+    const user = req.user;
+
+    if (id !== user._id && !user.isAdmin)
+      return handleError(
+        res,
+        403,
+        "Authorization Error: You must be an admin type user or the registered user to delete this user"
+      );
+
+    const userDeleted = await deleteUser(id);
+    return res.send(userDeleted);
   } catch (error) {
     return handleError(res, error.status || 500, error.message);
   }
